# Proposal: optimization of the json parser