This is a proposal to speed-up parsing of strings in a json-file.
I'm not sure that this is an issue for Flalbuffers because json is not the main goal of the library.
Benchmark result (strlen/before/after), time in nanoseconds per string in the json array:
   0  --  900ns  -- 135ns
 10  --  930ns  -- 170ns
 50  -- 1200ns -- 300ns
100 -- 1340ns -- 450ns
200 -- 1600ns -- 720ns
Most of the time (900-135=750ns) spended by uoffset->string->uoffset pattern.

@aardappel what do you think about this optimization?