@aardappel 2019-05-06
Yes, I would like the JSON parser to be really fast, and in particular, memory efficient. That was a goal at the start (when I was doing the benchmarks on the site), but may have slipped since then. So yes, lets optimize it!

Also, the trick of storing everything inside the `Value` string is terrible, and should be fixed.

That said, your solution is maybe a bit too hacky.. I'd prefer to refactor `Value` a bit more neatly. Some ideas:

- Simple: A `uint64_t` or whatever to act as storage of any scalars, along with a type field. Or maybe even a C union, such that no string is allocated if its never used.
- Crazy: most `Value` are used as temporary storage, so maybe have a dedicated buffer/stack where values can be pushed without allocation, like a `vector<uint8_t>` where arbitrary type/len values are appended? Or maybe we should use flexbuffers.h for this purpose, since it already is able to do these things.